product

+++ src/main/java/com/lyhour/developer/phoneshop_studying/repository/BrandRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 import com.lyhour.developer.phoneshop_studying.entity.Brand;
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer>, JpaSpecificationExecutor<Brand>{
+public interface BrandRepository extends JpaRepository<Brand, Long>, JpaSpecificationExecutor<Brand>{
 	List<Brand> findByNameIgnoreCase(String name);
 	List<Brand> findByNameLike(String name);
 }
